+++ backend/src/main/java/com/restaurant/dto/OrderItemDTO.java
@@ -1,17 +1,19 @@
 package com.restaurant.dto;
 
+import java.math.BigDecimal;
+
 public class OrderItemDTO {
     private String itemName;
     private Integer quantity;
-    private Double price;
+    private BigDecimal price;
     private String specialInstructions;
 
     public String getItemName() { return itemName; }
     public void setItemName(String itemName) { this.itemName = itemName; }
     public Integer getQuantity() { return quantity; }
     public void setQuantity(Integer quantity) { this.quantity = quantity; }
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
     public String getSpecialInstructions() { return specialInstructions; }
     public void setSpecialInstructions(String specialInstructions) { this.specialInstructions = specialInstructions; }
 }
